requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==2.1.3
scikit-learn==1.5.2
pydantic==2.8.2
httpx==0.28.1
pytest==8.4.2
pytest-cov==4.4.0
requests==2.31.0
qrcode==7.4.2
Pillow==10.4.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
APScheduler==3.10.4
google-cloud-bigquery==3.25.0
Flask==3.0.0

textblob==0.17.1
nltk==3.8.1

reportlab==4.0.7
qrcode==7.4.2
Pillow==10.1.0

